# fix pacing rate