Fixes #1854.
The pacing rate needs to be calculated for the next packet to be sent, not for the next byte to be sent.

## Benchmark Results

before:
```
+-----------------+-------------+--------------+
|                 |     TCP     |     QUIC     |
+-----------------+-------------+--------------+
| direct transfer | 0.02 ± 0.00 |  0.27 ± 0.01 |
| 5ms RTT         | 0.10 ± 0.00 |  0.44 ± 0.10 |
| 10ms RTT        | 0.20 ± 0.00 |  0.63 ± 0.16 |
| 25ms RTT        | 0.49 ± 0.00 |  3.39 ± 1.79 |
| 50ms RTT        | 0.82 ± 0.00 |  3.27 ± 1.18 |
| 100ms RTT       | 1.60 ± 0.03 | 14.09 ± 5.64 |
+-----------------+-------------+--------------+
Based on 3 samples (25 MB).
```

after:
```
+-----------------+-------------+-------------+
|                 |     TCP     |    QUIC     |
+-----------------+-------------+-------------+
| direct transfer | 0.02 ± 0.00 | 0.33 ± 0.02 |
| 5ms RTT         | 0.11 ± 0.00 | 0.49 ± 0.04 |
| 10ms RTT        | 0.22 ± 0.00 | 0.50 ± 0.04 |
| 25ms RTT        | 0.56 ± 0.01 | 1.03 ± 0.08 |
| 50ms RTT        | 1.08 ± 0.00 | 2.09 ± 0.06 |
| 100ms RTT       | 2.14 ± 0.00 | 5.40 ± 0.81 |
+-----------------+-------------+-------------+
Based on 3 samples (25 MB).
```

## Trace

See #1854 for the trace before this fix.
This is how a trace looks with this fix applied:
<img width="1457" alt="Screen Shot 2019-04-07 at 13 16 38" src="https://user-images.githubusercontent.com/1478487/55678597-6be8c100-5937-11e9-99ee-c362d6419207.png">